# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Location {
  latitude: number;
  longitude: number;
  timezone: string;
}

export type HtsiAssessmentRiskLevel = typeof HtsiAssessmentRiskLevel[keyof typeof HtsiAssessmentRiskLevel];


export const HtsiAssessmentRiskLevel = {
  LOW: 'LOW',
  MODERATE: 'MODERATE',
  HIGH: 'HIGH',
  VERY_HIGH: 'VERY_HIGH',
  EXTREME: 'EXTREME',
} as const;

export interface HtsiAssessment {
  value: number;
  risk_level: HtsiAssessmentRiskLevel;
  risk_color: string;
  description: string;
}

export interface WeatherSnapshot {
  temperature_c: number;
  relative_humidity_pct: number;
  wind_speed_ms: number;
  pressure_hpa: number;
  solar_radiation_wm2: number;
  cloud_cover_pct: number;
  weather_code: number;
  weather_label: string;
}

export interface WeatherCurrentResponse {
  location: Location;
  timestamp: string;
  source: string;
  calculation: string;
  weather: WeatherSnapshot;
  htsi: HtsiAssessment;
}

export interface ForecastPoint {
  timestamp: string;
  weather: WeatherSnapshot;
  htsi: HtsiAssessment;
}

export interface ForecastWarning {
  active: boolean;
  title: string;
  message: string;
  action: string;
  peak_value: number;
  peak_timestamp: string;
  peak_window: string;
}

export type WeatherForecastResponseHeatwaveRisk = typeof WeatherForecastResponseHeatwaveRisk[keyof typeof WeatherForecastResponseHeatwaveRisk];


export const WeatherForecastResponseHeatwaveRisk = {
  LOW: 'LOW',
  MODERATE: 'MODERATE',
  HIGH: 'HIGH',
} as const;

export interface WeatherForecastResponse {
  location: Location;
  generated_at: string;
  source: string;
  calculation: string;
  points: ForecastPoint[];
  peak: ForecastPoint;
  warning: ForecastWarning;
  heatwave_risk: WeatherForecastResponseHeatwaveRisk;
}

/**
 * Latitude in decimal degrees
 */
export type LatitudeParameter = number;

/**
 * Longitude in decimal degrees
 */
export type LongitudeParameter = number;

export type GetWeatherCurrentParams = {
/**
 * Latitude in decimal degrees
 * @minimum -90
 * @maximum 90
 */
lat: LatitudeParameter;
/**
 * Longitude in decimal degrees
 * @minimum -180
 * @maximum 180
 */
lon: LongitudeParameter;
};

export type GetWeatherForecastParams = {
/**
 * Latitude in decimal degrees
 * @minimum -90
 * @maximum 90
 */
lat: LatitudeParameter;
/**
 * Longitude in decimal degrees
 * @minimum -180
 * @maximum 180
 */
lon: LongitudeParameter;
};