# lib/api-zod/src/generated/types/forecastPoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HtsiAssessment } from './htsiAssessment';
import type { WeatherSnapshot } from './weatherSnapshot';

export interface ForecastPoint {
  timestamp: string;
  weather: WeatherSnapshot;
  htsi: HtsiAssessment;
}