# lib/api-zod/src/generated/types/forecastWarning.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ForecastWarning {
  active: boolean;
  title: string;
  message: string;
  action: string;
  peak_value: number;
  peak_timestamp: string;
  peak_window: string;
}