# lib/api-zod/src/generated/types/htsiAssessment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HtsiAssessmentRiskLevel } from './htsiAssessmentRiskLevel';

export interface HtsiAssessment {
  value: number;
  risk_level: HtsiAssessmentRiskLevel;
  risk_color: string;
  description: string;
}