# lib/api-zod/src/generated/types/weatherCurrentResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HtsiAssessment } from './htsiAssessment';
import type { Location } from './location';
import type { WeatherSnapshot } from './weatherSnapshot';

export interface WeatherCurrentResponse {
  location: Location;
  timestamp: string;
  source: string;
  calculation: string;
  weather: WeatherSnapshot;
  htsi: HtsiAssessment;
}