# lib/api-zod/src/generated/types/weatherForecastResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ForecastPoint } from './forecastPoint';
import type { ForecastWarning } from './forecastWarning';
import type { Location } from './location';
import type { WeatherForecastResponseHeatwaveRisk } from './weatherForecastResponseHeatwaveRisk';

export interface WeatherForecastResponse {
  location: Location;
  generated_at: string;
  source: string;
  calculation: string;
  points: ForecastPoint[];
  peak: ForecastPoint;
  warning: ForecastWarning;
  heatwave_risk: WeatherForecastResponseHeatwaveRisk;
}