# lib/api-zod/src/generated/types/weatherSnapshot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeatherSnapshot {
  temperature_c: number;
  relative_humidity_pct: number;
  wind_speed_ms: number;
  pressure_hpa: number;
  solar_radiation_wm2: number;
  cloud_cover_pct: number;
  weather_code: number;
  weather_label: string;
}